+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+
+### Added
+
+* hidden fields for custom fields which submit project ids of projects having
+  the respective custom field enabled but have no project type assigned
+
 ## 4.0.4 - 2021-05-29
 
 ### Deleted